[style] UserInfo는 User로 User는 Login으로 Codable 객체 이름 변경함
