+++ iOS/SignupApp/SignupApp/Models/Login.swift
@@ -1,0 +1,14 @@
+//
+//  User.swift
+//  SignupApp
+//
+//  Created by kimdo2297 on 2020/03/29.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+
+import Foundation
+
+struct Login: Codable {
+    let userId: String
+    let password: String
+}
