+++ iOS/SignupApp/SignupApp/Models/User.swift
@@ -11,4 +11,10 @@ import Foundation
 struct User: Codable {
     let userId: String
     let password: String
+    let name: String
+    let birth = "2020-03-24"
+    let gender = "male"
+    let email = "test@gmail.com"
+    let phone = "01012345678"
+    let interests = "영화감상, 자전거"
 }
